Extricate common library functions
